moved tests in memory-cache-store-size.js to memory-cache-store-tests.js
